- ErrorCode.java: keep BOOKING_NOT_COMPLETED + INSUFFICIENT_PAYMENT_AMOUNT alongside dev's new error codes
- BookingInvoiceRepository.java: merge methods from both branches (existsByBookingId, findByBookingId, findByBooking_Id)

+++ src/main/java/com/swp/autocarwash/common/contract/customer/VehicleContract.java
@@ -15,12 +15,12 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class VehicleContract {
-    private Integer id;
+    private Long id;
     private String licensePlate;
     private String brandName;
-    private Integer customerId;
+    private Long customerId;
 
-    public VehicleContract(Integer id, String licensePlate, String brandName) {
+    public VehicleContract(Long id, String licensePlate, String brandName) {
         this.id = id;
         this.licensePlate = licensePlate;
         this.brandName = brandName;
